Rename Cardano.CLI in Cardano.Wallet.Application.CLI
