+++ lib/exe/cardano-wallet-exe.cabal
@@ -112,4 +112,4 @@ library
   exposed-modules:
 
     Cardano.Wallet.Application
-    Cardano.CLI
+    Cardano.Wallet.Application.CLI
