+++ lib/unit/cardano-wallet-unit.cabal
@@ -87,6 +87,7 @@ test-suite unit
     , cardano-slotting
     , cardano-wallet-api
     , cardano-wallet-application-extras
+    , cardano-wallet-exe
     , cardano-wallet-launcher
     , cardano-wallet-network-layer
     , cardano-wallet-primitive
