+++ reports/security/2020-01-21.md
@@ -107,7 +107,7 @@ For example, when creating and logging checkpoint we have the following informat
 ```
 Hence, the name, creation date, delegation info of the wallet created earlier.
 
-Logging in lib/jormungandr/exe/cardano-wallet-jormungandr.hs, lib/wallet/src/Cardano/Wallet/Network.hs and lib/cli/src/Cardano/CLI.hs does not contribute to revealing of API calls info at all.
+Logging in lib/jormungandr/exe/cardano-wallet-jormungandr.hs, lib/wallet/src/Cardano/Wallet/Network.hs and lib/cli/src/Cardano.Wallet.Application.CLI.hs does not contribute to revealing of API calls info at all.
 
 **Conclusions** : The details of API logging is presented above. Sensitive information is concealed. Direct logging from API (as default it is Info severity level) is not giving very basic information about call method, endpoint, the response code, and time needed to realize it. Probably, it would be good to have Debug as a default - if we want to help other peoples and have maximum information possible without compromising their confidential info.
